src/app/app.component.spec.ts
import { TestBed, async } from '@angular/core/testing';

import { HttpClientTestingModule } from '@angular/common/http/testing';
import { FormsModule } from '@angular/forms';

import { AppComponent } from './app.component';
import { Config } from './model/config';
import { Language } from './model/language';

describe('AppComponent', () => {
  beforeEach(async(() => {
    TestBed.configureTestingModule({
      declarations: [AppComponent],
      imports: [HttpClientTestingModule, FormsModule]
    }).compileComponents();
  }));

  it('should create the app', () => {
    const fixture = TestBed.createComponent(AppComponent);
    const app = fixture.debugElement.componentInstance;
    expect(app).toBeTruthy();
  });

  it('should init', async(() => {
    const fixture = TestBed.createComponent(AppComponent);
    const app = fixture.debugElement.componentInstance;
    app.ngOnInit();
    app.getLanguage('C');
    app.updateRepositories();
    app.clearFilters();
    expect(app).toBeTruthy();
  }));

  it('should configurate', async(() => {
    const fixture = TestBed.createComponent(AppComponent);
    const app = fixture.debugElement.componentInstance;
    const config = new Config();
    const language = new Language();
    language.name = 'C';
    config.languages = [language];
    app.configurate(config);
    expect(app).toBeTruthy();
  }));
});
